Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/charset-02-expected.txt


PASS UTF-8 module script with UTF-8 BOM
FAIL UTF-16 module script with UTF-16BE BOM assert_equals: Should result in compile error because of UTF-16BE BOM expected (undefined) undefined but got (string) "三村かな子"
FAIL UTF-16 module script with UTF-16LE BOM assert_equals: Should result in compile error because of UTF-16LE BOM expected (undefined) undefined but got (string) "三村かな子"
